backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
pydantic==2.13.1
pydantic-settings==2.13.1
PyJWT==2.12.1
mistralai==2.4.0
requests==2.33.1
httpx==0.28.1
python-dotenv==1.2.2
python-dateutil==2.9.0.post0
